[MIG][9.0] Migrate FDS modules to version 9.0
[PORT] l10n_ch_fds: remove journal_id. Journal is matched automatically by import.

[PORT] l10n_ch_fds: link to bank statement

[PORT] l10n_ch_fds: remove l10n_ch_account_statement_base_import dependency

[PORT] l10n_ch_fds: set installable

[PORT] catch exception when python dependencies don't exist

[PORT] l10n_ch_fds_upload_dd: pep8 and lint check fixes

[PORT] l10n_ch_fds_upload_sepa: pep8 and lint check fixes

[PORT] l10n_ch_fds_postfinance: pep8 and lint check fixes

[PORT] l10n_ch_fds_postfinance: lint check fixes

[PORT] l10n_ch_fds_postfinance: lint check fixes

[PORT] l10n_fds_postfinance: remove journal_id

Fix bugs in postfinance FDS modules

Improve Postfinance Files Usability

- Make view more user friendly by directly showing all Downloaded Files
- Add More Menu Options to perform actions on several files at a time
- Add Option to exclude files from the download and import process
- Add Cancel state to hide some non-imported files
- Make Import FDS Wizard more accessible

- Add Wand library in travis config

- Fix security access rules

- Find related bank account when parsing Postfinance statement

- Correct some mistakes in parser
- Add date of statement
- Add account number so that the journal can be found automatically

- Parse statement dates when possible

Fix view of postfinance fds file to allow changing import journal

FIX Pylint and Tests fixes #232

FIX Pylint warnings and Exception handling in import FDS

disable module for now

[PORT] Change menu parent

Disable host key checking

[PORT] remove journal_id

Migrate module l10n_ch_fds_upload_sepa

Fix import of files and add CRON for daily import

Fix tests for SEPA Upload and FIX Pylint errors

Add tests for FDS keys generation

Improve code for requested changes

[PORT] Module l10n_ch_fds_upload_dd
