+++ l10n_ch_fds_upload_sepa/views/fds_postfinance_account_sepa_view.xml
@@ -6,7 +6,7 @@
             <field name="model">fds.postfinance.account</field>
             <field name="inherit_id" ref="l10n_ch_fds_postfinance.view_fds_postfinance_account_form"/>
             <field name="arch" type="xml">
-                <page string="Upload History" position="inside">
+                <page name="upload">
                     <separator string="SEPA"/>
                     <field name="sepa_upload_ids">
                         <tree create="false" default_order="create_date desc,state,filename">
