+++ l10n_ch_fds_upload_sepa/views/payment_order_upload_sepa_wizard_view.xml
@@ -1,0 +1,37 @@
+<?xml version="1.0" encoding="utf-8"?>
+<openerp>
+    <record id="wizard_pain_001_view" model="ir.ui.view">
+        <field name="name">Create SEPA Payment - Upload extension</field>
+        <field name="model">payment.order.upload.sepa.wizard</field>
+        <field name="arch" type="xml">
+            <form edit="1">
+                <header>
+                    <button string="Upload to FDS"
+                            name="upload_generate_file_button"
+                            type="object"
+                            icon="gtk-network" states="upload" class="oe_highlight"/>
+                </header>
+                <field name="state" invisible="True"/>
+                <sheet>
+                    <group states="upload">
+                        <field name="fds_account_id" widget="selection" edit="False" required="1"/>
+                        <field name="fds_directory_id"
+                               domain="[('allow_upload_file', '=', True), ('fds_account_id','=',fds_account_id)]"
+                               options="{'create': false, 'create_edit': false}" required="1"/>
+                    </group>
+
+                    <group>
+                        <field name="filename" invisible="True"/>
+                        <field name="file_data" filename="filename"/>
+                    </group>
+
+                    <group states="finish">
+                        <p style="color:red">
+                            <strong>Upload successful</strong>
+                        </p>
+                    </group>
+                </sheet>
+            </form>
+        </field>
+    </record>
+</openerp>
